Don't delete non-singular active interfaces.

+++ vnet/lib/vnet/node_api/active_interface.rb
@@ -16,13 +16,15 @@ module Vnet::NodeApi
         singular = options[:singular]
 
         transaction {
-          # Assuming label == nil for now:
 
           active_models = model_class.where(interface_id: interface_id).all
           Celluloid::Logger.debug "active interfaces pre-prune: #{active_models.inspect}"
 
           old_model = prune_old(active_models, datapath_id)
-          Celluloid::Logger.debug "active interfaces post-prune: #{active_models.inspect}"
+          Celluloid::Logger.debug "active interfaces post-prune_old: #{active_models.inspect}"
+
+          old_model = prune_for_singular(active_models, datapath_id) if singular
+          Celluloid::Logger.debug "active interfaces post-prune_for_singular: #{active_models.inspect}"
 
           if old_model
             model = old_model
@@ -31,6 +33,8 @@ module Vnet::NodeApi
                       singular: singular)
             model.save_changes
             
+            Celluloid::Logger.debug "updated active interface: #{model.inspect}"
+
           else
             model = super(options)
             Celluloid::Logger.debug "created active interface: #{model.inspect}"
@@ -74,8 +78,28 @@ module Vnet::NodeApi
 
           # TODO: Catch exceptions.
 
+          # Only destroy active entries that are old.
+
           # Currently default to destroying all:
 
+          # next unless active_model.updated_at + timeout < current_time
+
+          # active_model.destroy
+        }
+
+        old_model
+      end
+
+      def prune_for_singular(active_models, ignore_datapath_id)
+        old_model = nil
+
+        active_models.delete_if { |active_model|
+          if active_model.datapath_id == ignore_datapath_id
+            old_model = active_model
+            next
+          end
+
+          # Currently we destroy everything.
           active_model.destroy
         }
